student ui update
