+++ src/components/student/StudentAccount.tsx
@@ -53,8 +53,7 @@ function StudentAccount() {
   }
 
   return (
-    <div>
-      <h2 className="text-2xl font-bold mb-6">Account</h2>
+    <div className="max-w-md p-4">
       <Card>
         <CardHeader>
           <CardTitle>Edit Profile</CardTitle>
